pyproject.toml
[project]
name = "gpt-commit"
version = "0.1.0"
description = "Generate Git commit messages using GPT or local LLMs"
authors = [{name = "Puneet Kakkar", email = "pkakkar996@gmail.com"}]
license = "MIT"
dependencies = [
    "click>=8.0.0",
    "openai>=1.0.0",
    "python-dotenv>=0.19.0",
    "requests>=2.31.0"
]
requires-python = ">=3.7"

[project.scripts]
gpt-commit = "gpt_commit.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["gpt_commit"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "flake8>=4.0.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=gpt_commit --cov-report=term-missing"
